refactored text from layout
